Using MATHipow to reduce casting needed - changed type to be compatible

+++ src/libsac2c/support/math_utils.h
@@ -2,7 +2,7 @@
 #define _MATH_UTILS_H_
 
 extern int MATHlcm (int x, int y);
-extern long MATHipow (int base, int exp);
+extern int MATHipow (int base, int exp);
 extern int MATHnumDigits (int number);
 
 extern int MATHmin (int x, int y);
